feat - back: Add auth to deleteProductController

+++ Backend/src/controllers/product.controller.js
@@ -122,8 +122,21 @@ const updateProductController = async (req, res) => {
 
 const deleteProductController = async (req, res) => {
   const { pid } = req.params;
+  const productReq = req.body;
+  const authHeader = req.headers.authorization;
+  const token = authHeader.split(" ")[1];
+  const product = await productService.getProductById(pid);
 
   try {
+    const decodedToken = jwt.verify(token, config.jwtSecret);
+    const { user } = decodedToken;
+    const { role, userId } = user;
+    const { owner } = product;
+
+    if (role === "premiun" && userId !== owner) {
+      throw new Error("You can only delete products created by the same user");
+    }
+
     await productService.deleteProduct(pid);
     res.status(200).json({
       message: "Content successfully deleted!",
